Added required functionality for GPU module in LIB

+++ source/rose/roselib/LIB_memory_utils.hh
@@ -5,6 +5,8 @@
 
 #include "LIB_utildefines.h"
 
+#include <memory>
+
 namespace rose {
 
 /**
@@ -260,8 +262,8 @@ inline constexpr bool is_same_any_v = (std::is_same_v<T, Args> || ...);
  * Inline buffers for small-object-optimization should be disabled by default for large objects.
  * Otherwise we might get large unexpected allocations on the stack.
  */
-inline constexpr int64_t default_inline_buffer_capacity(size_t element_size) {
-	return (int64_t(element_size) < 100) ? 4 : 0;
+inline constexpr size_t default_inline_buffer_capacity(size_t element_size) {
+	return (element_size < 100) ? 4 : 0;
 }
 
 /**
